+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+* Enhanced the `hexapdf images` command to also show information on PPI (pixels
+  per inch) and size
 * Enhanced the `hexapdf inspect` command with an interactive mode, structure
   output and object search
 * Output of validation problem messages for `hexapdf` command to include more